chore: reach parity with upstream

+++ distributions/axoflow-otel-collector/preremove.sh
@@ -14,7 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-if command -v systemctl >/dev/null 2>&1; then
-    systemctl stop axoflow-otel-collector.service
-    systemctl disable axoflow-otel-collector.service
+if [ "$1" != "1" ]; then
+    if command -v systemctl >/dev/null 2>&1; then
+        systemctl stop axoflow-otel-collector.service
+        systemctl disable axoflow-otel-collector.service
+    fi
 fi
